Only install git if not installed.

+++ rachel.py
@@ -28,7 +28,8 @@ def die(d):
 	sys.exit(1)
 
 # Update and upgrade OS
-sudo("apt-get install -y git") or die("Unable to install Git.")
+if not (exists('/usr/bin/git') or exists('/usr/bin/local/git')):
+	sudo("apt-get install -y git") or die("Unable to install Git.")
 
 # Clone the repo.
 if exists("/tmp/rachel_installer"):
